+++ LineNumberGutter.cs
@@ -1,4 +1,7 @@
+using System;
 using System.ComponentModel;
+using System.Collections.Generic;
+using System.Drawing;
 using System.Windows.Forms;
 
 namespace DDF___Program_Language_Editor
@@ -9,6 +12,9 @@ namespace DDF___Program_Language_Editor
         private const int WmLeftButtonDown = 0x0201;
         private const int WmLeftButtonDoubleClick = 0x0203;
         private const int WmRightButtonDown = 0x0204;
+        private IReadOnlyList<int> displayedSourceLines = new int[0];
+
+        public event EventHandler<LineNumberClickEventArgs> LineClicked;
 
         public LineNumberGutter()
         {
@@ -23,20 +29,43 @@ namespace DDF___Program_Language_Editor
         [DesignerSerializationVisibility(DesignerSerializationVisibility.Hidden)]
         public Control TargetControl { get; set; }
 
+        public void SetDisplayedSourceLines(IReadOnlyList<int> sourceLines)
+        {
+            displayedSourceLines = sourceLines ?? new int[0];
+        }
+
         protected override void WndProc(ref Message message)
         {
-            if (message.Msg == WmSetFocus || message.Msg == WmLeftButtonDown ||
+            if (message.Msg == WmLeftButtonDown)
+            {
+                int y = unchecked((short)(((long)message.LParam >> 16) & 0xffff));
+                int character = GetCharIndexFromPosition(new Point(1, Math.Max(0, y)));
+                int displayedLine = GetLineFromCharIndex(character);
+                if (displayedLine >= 0 && displayedLine < displayedSourceLines.Count && displayedSourceLines[displayedLine] > 0)
+                    LineClicked?.Invoke(this, new LineNumberClickEventArgs(displayedSourceLines[displayedLine]));
+                FocusTarget();
+                return;
+            }
+
+            if (message.Msg == WmSetFocus ||
                 message.Msg == WmLeftButtonDoubleClick || message.Msg == WmRightButtonDown)
             {
-                if (TargetControl != null && !TargetControl.IsDisposed)
-                {
-                    TargetControl.Focus();
-                }
-
+                FocusTarget();
                 return;
             }
 
             base.WndProc(ref message);
         }
+
+        private void FocusTarget()
+        {
+            if (TargetControl != null && !TargetControl.IsDisposed) TargetControl.Focus();
+        }
+    }
+
+    internal sealed class LineNumberClickEventArgs : EventArgs
+    {
+        public LineNumberClickEventArgs(int sourceLine) { SourceLine = sourceLine; }
+        public int SourceLine { get; }
     }
 }